[feature] make ntpd_leapfile optional (#25)
* [feature] make ntpd_leapfile optional

fixes #24

* bump retries

"ntpd sync reaches all servers" failed on default-ubuntu-1404-amd64

* bump retry again

without-leapfile-ubuntu-1404-amd64 still fails to sync

* [documentation] document the new behaviour

* [bugfix] do not bailout when any upstream is INIT

the previous example bails out any one upstream is in INIT state, which
means the upstream is not working or reachable. the external resources
used in the test are not supposed to be available all the time. if not
all upstreams are in INIT state, fine.

* remove a duplicate

* sync with default_spec.rb